Add com.docker.slirp: a userspace TCP/IP proxy based on Mirage
This daemon listens on a Unix domain socket and speaks the vmnetd
protocol to xhyve. When the VM sends an ethernet frame it travels:

- through the Linux virtio-net driver in the VM
- to the xhyve virtio-ipc "hardware" in xhyve
- over the Unix domain socket via the vmnetd protocol
- into the Mirage TCP/IP stack
- into a regular Unix SOCK_DGRAM or SOCK_STREAM socket

The proxy will:

- assign the VM an IP with DHCP
- parse incoming DNS UDP packets and resolve them locally on the
  host (no direct forwarding to a DNS server)
- on receipt of a TCP SYN, issue a socket `connect` and
  - on `ECONNREFUSED`: send RST
  - on accept: send SYN ACK
- on receipt of other UDP, it will open up a NAT-style hole
  for 60s to a SOCK_DGRAM

This all means that:

- no OS X kernel bridges or NATs or routes are involved, and
  therefore there is nothing for VPNs to break
- all connections appear to come from sockets within the
  `com.docker.slirp` process on OS X, which should be routed
  down the VPN like all other applications' network connections

The performance as measured by `nc -l` on the host and
`cat 1GiB | nc` on the VM is about 1 Gbit/sec.

The main limitation is that ICMP doesn't work (actually responses
get sent from the host by accident).

Note further that no root access is needed for this kind of
networking, it is sandbox-friendly.

Signed-off-by: David Scott <dave.scott@docker.com>