com.docker.slirp: drop packets with bad source IPs on the link to Moby
Occasionally when `apt-get` installing packages as part of the Moby
build (see #821), the VM transmits an IP packet with an internal
172.17.0.2 source address, rather than a NATted source address (in this
case 10.0.0.2). This patch drops these packets and logs the
source, destination and protocol.

The packets in question are heading to 176.9.184.93:80, one of the
current httpredir.debian.org addresses.

Fixes #837

Signed-off-by: David Scott <dave.scott@docker.com>