com.docker.slirp: allow the internal IP to be customised
On the point-to-point ethernet link between com.docker.slirp and
the VM we have to run IPv4, so we need to allocate a small IPv4
network with 2 IPv4 addresses.

We default to private addresses from RFC1918: 10.0.0.1 and
10.0.0.2, but some enterprises use these addresses for internal
services. The effect would be to hide these services from the
VM, possibly breaking functionality.

This patch allows the private addresses to be changed via the
following database keys:

com.docker.driver.amd64-linux/slirp/host: for the Docker.app end
com.docker.driver.amd64-linux/slirp/docker: for the VM end

The netmask will be generated from these 2 IPs, keeping it as
small as possible (although to work around a bug in the DHCP
server it needs to include 2 more IP addresses)

Signed-off-by: David Scott <dave.scott@docker.com>