com.docker.slirp: switch default IP net to 169.254.0.0
Nomatter which IPs we choose for these internal links, they might
clash with some setup an enterprise is using. Many places will route
RFC 1918 addresses internally so if we use those then we end up
masking part of the address space. The symptom would be some IPs
would be visible from the host but not the VM or containers.

This patch switches us to the IPv4 autoconfiguration space instead,
which is less likely to be routed anywhere. Note other people do
this too, for example AWS places config files on a server on
169.254.169.254, and cloudstack uses this network for its system VM.

Unfortunately the default has to be encoded in 2 places because
we don't have an entity which starts early and manages the db
defaults. If xhyve starts first, it must know the defaults. If
slirp starts first, it also must know.

Signed-off-by: David Scott <dave.scott@docker.com>