docker-proxy: discover and rewrite ports in `docker ps`
If a port forward request specifies the host port, then the forward is
setup in the `start` codepath as before. If the host port is unspecified
then we delay opening the port until the user first queries for it
via `docker ps`, then we can observe the open port on Moby and create
a local tunnel to it.

Signed-off-by: David Scott <dave.scott@docker.com>