com.docker.slirp: avoid restarting unnecessarily on first start
Before this patch if the database was empty then `com.docker.slirp`
would consider the configuration to be unset (`None`) and supply its
own defaults. When `xhyve` populated the defaults later, `com.docker.slirp`
would think that the settings had changed (when they hadn't) and would
exit. This caused unnecessary service bouncing on first start.

This patch refactors the database watching code to accept `default`
arguments and to expose the `map` function over the stream of
values. Now `com.docker.slirp` can tell the database watching
code what the value is, and it won't be told of any spurious changes.

Signed-off-by: David Scott <dave.scott@docker.com>