com.docker.slirp: respect DNS changes on the host
Before this patch we cached a resolver which had parsed
`/etc/resolv.conf`. Unfortunately if the wireless network changed
then the `/etc/resolv.conf` would be rewritten but we wouldn't
notice.

This patch makes us re-parse `/etc/resolv.conf` on every request.
This is a bit heavyweight -- it would be better to cache the file
and receive events when it changes -- but we can improve that
later. At least the `/etc/resolv.conf` is very short and quick
to parse.

Fixes #1435

Signed-off-by: David Scott <dave.scott@docker.com>