com.docker.slirp: add an internal API to start/stop pcap capturing
This patch removes the command-line argument from com.docker.slirp
for pcap capturing because it was hard to use -- killing the daemon
would cause it to be restarted by the app.

This patch also adds new `start_capture` and `stop_capture` APIs to
the vmnet library which we can use later to dynamically start and stop
capturing. There's also a size limit which will be used to prevent
the local disk filling up by accident.

Signed-off-by: David Scott <dave.scott@docker.com>