com.docker.slirp: transmit packets with the intended MAC address
Before this patch we accidentally set the source MAC address on frames
sent by the server to the *client*'s MAC address. This patch ensures
that the MACs in use are defined in only one place and everyone
knows which MAC is the client's and which is the server's.

Note we should probably change the server's MAC since wireshark highlight's
it's an illegal address.

Signed-off-by: David Scott <dave.scott@docker.com>