ofs: make he active_list protocol concurrency-safe
Before this patch there was a shared `ctl` file within the filesystem
whose state was tied to the connection. This works ok when making
socket connections to the server, but is not ok when mounting in
Linux as separate processes can access the control file concurrently.

This patch changes the protocol, so every forward has a name. To
create a forward:

  mkdir name
  echo -n 'unix:/tmp/foo.socket:127.0.0.1:80' >> name/ctl
  cat name/ctl

To delete a forward which failed:

  rm name/ctl

To delete a forward which succeeded:

  rm name

To discover the forwarding rule associated with a name:

  cat name

Note that, as well as per-name control files, we nolonger encode the
forwarding rule in filenamea. Before this patch the string form
of the forwarding rule -- often containing path separators -- was
used as a filename in the protocol, which is obviously illegal.

Signed-off-by: David Scott <dave.scott@docker.com>