Shutdown port forwards on clunk and disconnect
Previously we would tear down port forwards (and approved volumes) by
`rm` the magic file. Unfortunately if the VM crashes or if the `docker-proxy`
crashes this won't happen. Instead we rely on the client keeping their
fd (or Fid) open, and we tear down the forward on clunk or disconnect.

This covers the following cases:

- if the `docker-proxy` crashes, the kernel will clunk the Fid and the
  proxy will be terminated.
- if the OS crashes then xhyve will exit, the connection on the host side
  will disconnect

This requires the related Moby change to be merged first.

Fixes #2475

Signed-off-by: David Scott <dave.scott@docker.com>