Add database key allowed-bind-address
If this key is absent (the default), then there is no restriction on
the external addresses ports may be bound on.

If this key is present, then it is parsed as a comma-separated list of
IP addresses. An external bind will be allowed if it matches any of
the given IP addresses.

Needed by #2091

Signed-off-by: David Scott <dave.scott@docker.com>