com.docker.slirp: functorise the slirp stack over the network
Before this patch the slirp stack assumed that it could use the Vmnet
module to talk ethernet to the VM. The Vmnet module itself expects
a Unix domain socket and a particular protocol. This patch functorises
the slirp stack so that a more Windows-specific network implementation
can be provided which may use a different protocol (and certainly
won't involve any Unix domain sockets).

Signed-off-by: David Scott <dave.scott@docker.com>