hostnet: update the opam metadata for Windows build
Since the convenient Windows build starts with fdopen/opam-repository
this patch bumps the version number of upstream packages we forked to
a very high and obvious number (999) so we don't accidentally flip back
to upstream when they release a new version.

Depend on hashcons.1.0.1 since the later version depends on conf-pkg-config
which isn't available.

Signed-off-by: David Scott <dave.scott@docker.com>