com.docker.slirp: connect to a single well-known port forwarding port
Previously we mapped exposed ports on the host (localhost) to vsock ports
in the range 0x10000-0x20000 for TCP and 0x20000-0x30000 for UDP. Unfortunately
this approach doesn't work for Hyper-V sockets since we have to manually
register GUIDs in the Windows registry. This patch switches to a different
scheme where there is one well-known port forwarding port, we connect to it
and write a header describing the ultimate destination, and the proxy redirects
us. This should probably be replaced with a SOCKs proxy.

Signed-off-by: David Scott <dave.scott@docker.com>