hostnet: Round-Robin DNS retries through the available upstream DNS servers
If a DNS request times out then applications will retry with the next available
server. However we only expose one DNS endpoint to the guest (ours) even if we
might have multiple upstreams, this means that DNS looksup can fail due to
transient issues where they would succeed natively.

To fix this keep a cache of outstanding DNS transaction ids and if we see what
looks like a retry (i.e. a cache hit) send it to the next server instead of the
first one again. Entries are removed from the cache either when we see a reply
or after 1 minute.

If we can't parse the DNS packet at all then we always forward it to the first
server.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>