hostnet: include a single definition of `Port.t`
Previously each frontend exposed a slightly different type: the Mac
would use a virtio-vsock and Windows was going to use a AF_HYPERV
GUID. However now that the protocol has changed so the destination port
is now requested in a SOCKS-like protocol, both Mac and Windows use
a <protocol:IP:port> triple when setting up the forward. This means
we can coalesce the `Port.t` types into one.

Signed-off-by: David Scott <dave.scott@docker.com>