proto-vmnet: use a buffered Channel rather than raw FLOW
Instead of calling `FLOW.read` and `FLOW.write` directly which map onto
`read(2)` and `write(2)`, layer a buffered Channel on top. This will
allow us to reduce the number of times we make syscalls, for example
we could read several packets in one go if the underlying `FLOW` has
a big enough read buffer.

Signed-off-by: David Scott <dave.scott@docker.com>