unix: increase read buffer size from 1024 to 65536 bytes
Now that we are using a buffered Channel, we can read more bytes at
once from Unix file descriptors. This should reduce the rate at which
we make expensive syscalls.

With a simple iperf test sending from a container and receiving on
the host, this boosts throughput from 1.01 Gbits/sec to 1.33 Gbits/sec
on the Mac.

Signed-off-by: David Scott <dave.scott@docker.com>