dns: share the DNS NAT code
Rather than use bare `Lwt_unix.send`, `Lwt_unix.recvfrom` this patch
shares the existing code in `Socket.Datagram`. This will make a future
patch to use libuv smaller.

The main difference between the two `recvfrom` uses is that the previous
DNS one would always timeout after 5s, whereas the general UDP NAT
(activated for example if the VM sends DNS to 8.8.8.8) will be listening
for 60s. So if a response is delayed > 5s we might receive it now.
Presumably the same thing happens on the Internet anyway as there is
no `Lwt.cancel` for transmitted UDP packets.

Signed-off-by: David Scott <dave.scott@docker.com>