hostnet: fix TCP DNS (required when UDP responses are truncated)
TCP connections sent to our IP are normally forwarded to the host,
except for port 53 (DNS) which is forwarded to an upstream DNS
server.

Previously we checked whether a connection was "for us" by comparing
the IP destination to our IP. If a connection was for port 53 we would
then rewrite the IP destination but we failed to reconsider whether
this new IP was ours or not. This meant that we always forwarded TCP
on 53 to localhost by mistake.

This patch makes the "is connection for us" decision into a
function `for_us` and re-evaluates it when the destination IP may have
changed.

This fixes a bug introduced in a56356fd427870f7f3679b8dc9db6f403ea394db

Signed-off-by: David Scott <dave.scott@docker.com>