Add a secondary DNS forwarding IP
In #38 I attempted to address the problem of flakey upstream DNS by
round-robinning DNS retries (detected via the DNS ID field), replicating the
fallback behaviour a client which has been given multiple entries in
resolv.conf despite only providing then with a single entry.

However this turns out to have been insufficient because not all client side
resolvers reuse the DNS ID on retry, so we do not see those retries as such and
send them to the primary upstream DNS a second time.

To address this create a second IP end point (on the same underlying device, so
essentially an IP alias)  and expose it to the client (via DHCP) as a second
DNS server which forwards to the secondary upstream DNS server and revert most
of #38 such that for the existing local IP we always forward to the primary
upstream DNS server (if the host only has a single DNS server configured then
both end points will go to the that one).

Third and subsequent host DNS servers are ignored.

This has only been tested on Mac, but given an implementation of Resolv_conf
which returns 2 (or more) DNS end points this should work for Windows too.

This second IP address is bolted onto the side in what appears to be a slightly
odd way. However I think it is consistent with the direction (AIUI) we intend
to be moving which is to remove the use of `Tcpip_stack_direct` (since we
aren't really a TCP/IP "endpoint") in favour of a lower level approach (based
on the newer Marshal/Unmarshal interfaces in mirage-tcpip rather than even more
open coding). Currently mirage-tcpip does not support aliases (second IPs) on a
device, so going the route of creating a second full stack would have required
an entire extra network device be exposed to the client.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>