Move Vmnet.ml{,i} into hostnet
This small module isn't shared by anyone else and is really implementing
a private protocol shared between hyperkit and vpnkit. Ideally in the
future we would replace this with something standard (PPP?)

This patch simplifies the build a little by removing one of the dependencies
where the overhead of management (in terms of build and package metadata and
in terms of extra developer time taken to manage it) is greater than the
actual logic in the module.

Signed-off-by: David Scott <dave.scott@docker.com>