uwt: when `write` fails with `EPIPE`, this is a "normal" `Eof
When a peer process calls `close` and we try to `write`, the call
will fail with `EPIPE`. Since we're often proxying arbitrary protcols,
this can be completely legitimate behaviour. Previously we would emit
a scary-looking error log, making people think there's a bug in the
proxy.

This patch suppresses the log line.

Signed-off-by: David Scott <dave.scott@docker.com>