Filter IPv4 DNS resolvers in the hostnet code
We wish the IPv4-only filter to be applied nomatter where the DNS servers
are read from including

- the `/etc/resolv.conf` file
- the database keys
- the command-line

Since the limited code is inside the core of the hostnet library it makes
sense to filter there. When the core is upgraded to support IPv6 we can
remove the filter at the same time.

Signed-off-by: David Scott <dave.scott@docker.com>