Support TCP DNS on the secondary IP
Previously the UDP DNS code would choose a host server depending on
which IP the VM chose to use, but the TCP DNS code would always forward
to the primary server. This patch shares the server choosing code between
UDP and TCP, so that TCP DNS to the secondary IP gets forwarded to the
secondary server.

Fixes #96

Signed-off-by: David Scott <dave.scott@docker.com>