Generalise to muliple DNS IPs, use 8 IP addresses by default
We expose multiple DNS server IPs to the client to make use of
the client's built-in fallback strategy in case a server does
not respond or a name is not found. Some systems (especially
with VPNs configured) have large numbers of resolvers configured,
so we need to support a large number of resolvers here too.

This patch gives the VM 8 DNS IPs, but more may be configured
via the database. The existing primary/secondary selection
algorithm is unchanged, so all these IPs are currently mapped
onto the first 2 host resolvers.

Note this patch removes one user of the Tcpip_stack_direct: instead
of having some DNS logic in `listen_udpv4` and others in a
packet-parsing callback, all of the code is now in the packet-parsing
callback.

Signed-off-by: David Scott <dave.scott@docker.com>