Use fewer UDP NAT rules
In the very first version we associated NAT rules by

  (dst_ip, dst_port)

which was probably a mistake brought on by using `src` and `dst`
both ways round in slirp.ml -- a confusion which we need to fix.
This meant that requests from different containers to the same
server (e.g. DNS) would conflict over the rule and only one would
win.

We then switched to a bigger tuple for safety:

  (src_ip, src_port, dst_ip, dst_port)

Note the listening sockets on the physical side were always listening
for traffic from everywhere (i.e. full cone NAT) but we ended up creating
lots of duplicated sockets -- one per destination -- all of which
would forward back to the internal address.

The duplicated sockets wasted resources and it meant that each one
was timed-out separately: a heartbeat to one destination address
wouldn't keep all the rules alive, meaning another external client
could discover the their specific rule had expired even though another
host is still successfully heartbeating.

This patch switches to a NAT rule per source address

  (src_ip, src_port)

This was probably the configuration we wanted in the first place.
This reduces the number of sockets we will allocate, and makes a
NAT traversal easier to manage.

Signed-off-by: David Scott <dave.scott@docker.com>