test: verify that 2 source ports create 2 NAT rules
In the past there was a bug where packets to the same destination
server (ip, port) used the same NAT rule. This test case generalises
the UDP echo server into a "chat" server, and checks that a message
from a second local port successfully triggers replies to both the
second local port and the first one.

Signed-off-by: David Scott <dave.scott@docker.com>