test: add NAT traversal case
This case sends UDP from the virtual address to a real physical
address to set up a NAT rule. It then uses another socket with a
different physical address and sends traffic to the external
address used by the NAT. Hopefully this traffic also traverses the
NAT and hits the packet counter on the inside.

Signed-off-by: David Scott <dave.scott@docker.com>