DNS: serve entries read from /etc/hosts
When we receive an A or AAAA DNS request we first look in our cache
of /etc/hosts entries. If it's not in our cache then forward it to
the upstream server as normal.

This allows users to edit /etc/hosts on the Mac/Windows host and then
resolve these names from containers.

Signed-off-by: David Scott <dave.scott@docker.com>