Reduce the amount of DHCP log spam
Now that we're using a short lease we generate a lot of logs: one
line per BOOTREQUEST and BOOTREPLY every half (?) lease interval.

This patch logs the first BOOTREQUEST and first BOOTREPLY as `info`
rather than `debug` since these are genuinely useful things to see
in the logs. Subsequent BOOTREQUEST and BOOTREPLY messages are
suppressed. Other message types are logged as `info`.

Signed-off-by: David Scott <dave.scott@docker.com>