Use upstream APIs
Previously we used a modified version of the Mirage `Tcpip_stack_direct`
which

- relied on the bug where TCP SYNs would be ACKed even if not addressed
  directly to us (e.g. if we're acting as an IP router)
- added a flow `Accept`/`Reject` API to avoid having to register
  65536 hashtable entries, one per port.

This PR changes the wiring as follows:

- incoming packets are switched based on destination IPv4 address to
  separate custom stacks
- each custom stack uses the regular interfaces, has a single IPv4
  address. We don't rely on the TCP SYN accept bug any more.
- each custom stack has a custom `Netif.listen` callback which
  pattern-matches on incoming packets and intercepts TCP SYNs.
  Before the SYN is forwarded, the stack will issue a `connect`.
- packets which don't match a switch rule are sent to a default
  callback which sets up a new stack and configures a switch port (c.f.
  the openvswitch control plane)

As a readability benefit, we now use pattern-matching everywhere rather
than a mixture of `add_listener`-style callbacks and pattern-matching.

All the new state is visible over the 9P debug interface.

It should be possible to compile this code against more recent versions
of the Mirage TCP/IP stack, but this PR doesn't attempt to do so.

Signed-off-by: David Scott <dave.scott@docker.com>