DNS: capture the host-side UDP datagrams too
Previously the packet capture logic would only see the VM-side datagrams
which made it unnecessarily difficult to debug DNS forwarding problems.
This patch synthesizes UDP datagrams corresponding to the `send` and `recv`
on the host and injects them into the capture module.

The captured traces now include both the host<->VM and the host<->upstream
transactions.

Signed-off-by: David Scott <dave.scott@docker.com>