Fix the recording of external DNS requests
This fixes a bug where the recording function was actually a partial
application. Note in order to get the `Recorder.t` value to the
DNS forwarding callback we now use a global variable. This is because
the DNS forwarder is re-created by the database watching code in
`Slirp.create` while the recorder is only available later in `Slirp.connect`.
Perhaps the `create` and `connect` split should be refactored.

Signed-off-by: David Scott <dave.scott@docker.com>