Rewrite packets sent by DNS forwarder to the slirp IP
Previously the DNS forwarder would call us back with copies of the
DNS packets it has sent and received. For the IP address it uses
`getsockname` which will return the bind address i.e. 0.0.0.0.
This patch rewrites this to be the virtual IP used by the host on
the slirp link, so the resulting .pcap traces look plausible.

Signed-off-by: David Scott <dave.scott@docker.com>