Clarify exactly which modules need `read_into`
Previously we required all `FLOW_CLIENT` instances to support
`read_into` which had the side-effect of requiring an implementation
for Udp which didn't make any sense so had to return a runtime error.

This patch creates a `READ_INTO` mixin signature and mixes it in only
where it is actually needed, so we can remove the bogus `Udp` functions.

Signed-off-by: David Scott <dave.scott@docker.com>