Move the Cstructs module to the top-level
This module provides a subset of the interface provided by Cstruct
except the `type t = Cstruct.t list`. This allows easier parsing of
packets which have been read in several chunks (e.g. from a CHANNEL)

Probably this should be replaced with some more fully-featured
parser such as angstrom (?) when that becomes available.

Signed-off-by: David Scott <dave.scott@docker.com>