Frame.parse now takes a Cstruct.t list
Most of our packets are contained within a contiguous `Cstruct.t` but
some (notably those generated by the stack via `writev`) are in
`Cstruct.t list`. Due to the two different types we ended up with two
different parsers: one in `Frame` and one in `Match`.

This PR generalises the parser in `Frame` to parse `Cstruct.t list`s
so that we can drop one of the parsers in a future patch.

Signed-off-by: David Scott <dave.scott@docker.com>