Frame: add an explicit Unknown type
Previously we wouldn't distinguish between an obviously malformed packet
(e.g. truncated) and a packet that we don't understand. There are clearly
many types of packets that this parser doesn't understand (like IPv6!)
so represent these as an explicit `Unknown` type and return success from
the function.

This avoids debug spam for every IPv6 packet and avoids filling up the
bad.pcap with these false positives.

Signed-off-by: David Scott <dave.scott@docker.com>