test: check that UDP source ports are preserved
This test is based on the example from [docker/for-mac#926] where
we have 2 external echo servers and one internal client/server
socket. The internal socket sends to both echo servers and records
the source ports of the replies.

This triggers a bug where the initial order of transmission wins,
since the source port is captured at the time the forwarding socket
and shared NAT rule is created.

Signed-off-by: David Scott <dave.scott@docker.com>