Udp: fix packet overwriting bug in listen
Previously the `listen` loop allocated a buffer and called `recvfrom`
to read a UDP datagram. The flow callback was called asynchronously
with a reference to this buffer, so the next `recvfrom` could overwrite
the contents.

This patch uses a fresh buffer per `recvfrom` which avoids accidentally
overwriting a previous packet.

Signed-off-by: David Scott <dave.scott@docker.com>