De-duplicate the UDP NAT code
Previously the UDP NAT code was duplicated inside the Host_uwt and
Host_lwt_unix backends because it used some internal APIs. Afterwards
we ended up exposing the required APIs anyway (`sendto` and `recvfrom`)
so this code can be abstracted over the datagram sockets interface.

This patch reduces the amount of code that will need to be written
to fix the broken test in #139

Signed-off-by: David Scott <dave.scott@docker.com>