Uwt: remove an unnecessary `deregister_connection`
Previously we deregistered the connection and then called `Lwt.fail`.
Since we're in the middle of an `Lwt.catch`, control passes to the
exception handler which calls `deregister_connection` again. In most
cases this logs a benign error but it some cases we might deregister
some other connection which could cause the connection limiting to
underestimate the total number of active connections.

Signed-off-by: David Scott <dave.scott@docker.com>