Update dns-forward to v0.6.0
This includes an important bugfix: if a single server fails fast,
don't cancel all other resolution attempts. For example if we have
an IPv6 server configured but we don't have an IPv6 network on the
host then we expect the `sendto` to fail. We wish the remaining
IPv4 RPCs to continue and the overall resolution to fail only if
all threads have failed by the timeout.

Signed-off-by: David Scott <dave.scott@docker.com>