test: move the port forwarding tests to earlier in the test suite
The Windows CI tests are failing intermittently -- see #156. One
of the symptoms is that TCP/IP stack threads run and log, even though
this particular test doesn't use the TCP/IP stack -- it's testing the
9P port-opening control-plane.

This patch moves the forwarding tests earlier in the sequence so they
shouldn't overlap with Lwt threads from earlier tests of the stack.
Maybe this will help, maybe it won't.

Signed-off-by: David Scott <dave.scott@docker.com>