Disable -warn-error for warning 3: deprecated functions
Ideally we would keep this as -warn-error for developer builds but disable
for tagged release builds. This patch disables -warn-error for warning 3
for both cases: it's not worth too much work on the current build and
packaging system when we're likely to switch over to topkg soon.

This allows vpnkit to build with lwt.2.7.0.

Related to [ocaml/opam-repository#8195]
Related to [ocaml/opam-repository#8196]

Signed-off-by: David Scott <dave.scott@docker.com>