Add global arp table
A mutable IP/MAC-table is added to the config struct. Each new
slirp connection adds a new pair to this table. When
an ARP-request is received, a real ARP-table is created
based on the content of the global table. The client
connected in the other end should then be able to resolve
all slirp users via the global ARP.

Endpoints still have a local ARP table that just contains
the IP/MAC pairs used in the endpoint.

ARP-entries are not expired, even if the slirp connection
is terminated. This should probably be fixed if a lot of
VM restarts is expected.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>