If no external DNS servers configured, use getaddrinfo
This is part of a move towards using `getaddrinfo` for all lookups
by default. The advantage is that it should "just work" for the
common cases with VPNs via the normal host resolver. The
disadvantage is that it doesn't provide a means to query other
DNS information such as MX records. Either this will need to be
fixed in future, or perhaps such advanced use requires explicit
configuration of an upstream nameserver anyway.

Signed-off-by: David Scott <dave.scott@docker.com>