dnssd: recursively look up CNAMES
If we don't supply the `kDNSServiceFlagsReturnIntermediates` flag
and try to resolve a record that doesn't exist e.g.

  host -t MX dave.recoil.org

then the call will only return after a (long) timeout. The dnssd
binding supplies this flag to guarantee calls return quickly.

Unfortunately this has the side-effect that some names resolve only
to a CNAME, which causes clients to reject the response. In this patch
we fix this problem by recursively calling `Dnssd.query` to resolve
these CNAMES.

Signed-off-by: David Scott <dave.scott@docker.com>