DNS: disable mDNS
Currently a query for `foo.local` can use the host resolver and this
may involve real mDNS, which can be very slow. At the moment our current
client (linuxkit, previously known as Moby) has a domain name set to
`local` which causes lots of lookups of `foo.local` names which don't
exist, which each incur a ~5s timeout.

This patch returns `NXDomain` for all lookups of `*.local` for now.

Signed-off-by: David Scott <dave.scott@docker.com>