uwt: catch exceptions and continue around `close_wait`
The error log in [docker/for-mac#426] shows an unguarded failure within
an `Lwt.async` call, which terminates the thread and calls the
`async_exception_hook`. Since the failure happens in the middle of a stressful
run where the ports are bound near the beginning, it must be happening
somewhere inside the per-connection callback, rather than in the initial
setup with `listen`.

Looking inside the `Lwt.async` within the connection callback, the only
code which is not guarded with an exception handler is the `close_wait`
within the `finally` cleanup handler. This patch adds exception handlers
which log the error and continue, which is the best we can probably do
if `close_wait` fails.

Signed-off-by: David Scott <dave.scott@docker.com>