DNS: add optional builtin DNS names for the Host
Previously the only way to access ports exposed on the Host from the VM
was to know and hardcode the virtual IP `192.168.65.1` (although this
can be changed via the database).

This patch adds a command-line argument --host-names which accepts a
comma-separated list of DNS names which will be mapped to the current
host virtual IP.

The default name is

  vpnkit.host

but this can be overriden or extended.

Signed-off-by: David Scott <dave.scott@docker.com>