Add some weekly VPNKit development reports
These reports have been generated from GitHub activity, with some
hand editing in the style of the LinuxKit reports. They can be
browsed directly within the repository and cover DataKit and
immediately related repositories such as ocaml-git and 9p. Fixes
can be contributed directly to this repository.

For the brave who want to generate them for their own repositories,
the work-in-progress generator of the skeleton of these reports
is in https://github.com/avsm/hesternus

Signed-off-by: Anil Madhavapeddy <anil@recoil.org>