Fix warning by not catching `Failure` exceptions
Previously we used `Failure` as a quick means of jumping out of an
expression to the exception handler. However `Failure` is a special
exception where the string is only meant to be recorded, not matched.

This patch uses a first-class `Enoent` exception.

Note the `Failure "BADWALK"` case could never happen: no code will
throw this error.

Signed-off-by: David Scott <dave.scott@docker.com>