CI: avoid running the unit tests of cohttp
These tests appear to block on appveyor.

As a general rule we don't re-run the unit tests of all our dependencies
here, but we make a special exception for tcpip which we have heavily
modified. This patch installs and runs the tests for tcpip first, before
installing (without tests) all the other dependencies.

Signed-off-by: David Scott <dave@recoil.org>