test: check what happens after we hit the connection limit
We impose a connection limit to avoid hitting the system limits
(particularly the global limits on Mac). Previously the test would
check that we couldn't exceed the connection limit. This patch
extends the test to check that, when the limit is raised, we
can still make outgoing connections. This verifies that the act
of hitting the limit does not break the stack i.e. we should be
able to recover afterwards.

Signed-off-by: David Scott <dave.scott@docker.com>