add vpnkit-iptables-wrapper
This wrapper calls vpnkit-expose-port when swarm service ports are
opened by dockerd using iptables.

To be used up by dockerd it has to be in path before the regular
iptables. The wrapper expects /sbin/iptables to contain the original
binary.

This is a go version of the iptables-wrapper originally written in
OCaml:

https://github.com/linuxkit/linuxkit/blob/f27c3ff5eda58d76e49ce5129b3cb29e402be272/alpine/base/pinata-iptables/main.ml

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>