Dhcp: make the dynamic DHCP range configurable
When `bridge-connections` is enabled the server will reserve more IP
addresses for clients such as LinuxKit which connect to the vpnkit
network. Previously it was not possible to adjust fully the
range of addresses handed out, so if you had a clash you were out of
luck.

This patch plumbs through a new `slirp/highest-ip` setting so that
clashes can be worked around.

Related to [docker/for-mac#1803]

Signed-off-by: David Scott <dave.scott@docker.com>