test: add cases for half-close
This patch adds 2 simple test cases, where a server is run on localhost
and connected to via vpnkit via the Mirage TCP client. Both cases use
a simple request / response protocol where the connection is half-closed
after the request is written and before the response is written, to check
that the response is still relayed correctly.

The `test_mirage_half_close` performs the half-close via the Mirage stack;
the `test_host_half_close` performs the half-close via the Host stack.

Signed-off-by: David Scott <dave.scott@docker.com>