connect: return Error when we exceed the connection limit
Previously we would fail the thread with `Too_many_connections` but this
meant that the SYN was never responed to. With this change, we send explicit
RST packets to the caller.

This makes the `test_max_connections` test pass -- previously the TCP
`connect` call would resend SYNs for a long time.

Signed-off-by: David Scott <dave.scott@docker.com>