Make tests work in bridge mode
This swaps client/server macaddr in client_of_fd, so the server_macaddr
can be passed in instead of the client_macaddr. Previously packets
would be sent with the default client MAC address hardcoded in
Slirp_stack, but as the bridge now assigns the MAC it has to be dynamic
to pass the MAC filter (wrong src MAC is dropped).

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>