Add Preferred_ipv4 slirp command
This adds a new command "Preferred_ipv4" that allows the client to
request an IPv4 address when creating the connection to vpnkit. The
command will fail if the IP is already in use by a different UUID or the
UUID has previously been used with a different IP.

As the length of the Command packet had to be extended to make room for
the IP address, the protocol version is bumped to 2. This requires
updates in other projects that use VPNKit. If the client requests an
unsupported version (1), the server will respond with the supported
version and disconnect.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>